Add floating point support.

+++ src/lexer.rs
@@ -90,10 +90,10 @@ pub fn tokenize_line(line: &str, store_text: bool) -> Result<LineOfCode, String>
                     token_chars.insert(0, ch);
                     let token_str: String = token_chars.into_iter().collect();
 
-                    if i32::from_str(token_str.as_str()).is_ok() {
+                    if f64::from_str(token_str.as_str()).is_ok() {
                         tokens.push(TokenAndPos(
                             pos,
-                            token::Token::Number(i32::from_str(token_str.as_str()).unwrap()),
+                            token::Token::Number(f64::from_str(token_str.as_str()).unwrap()),
                         ));
                     } else {
                         let token = token::Token::token_for_string(token_str.as_str());
